requirements.txt
flask
flask-socketio
playwright
pyyaml
requests

